requirements.txt
# Core dependencies
numpy>=1.24.0
opencv-python>=4.8.0
mediapipe>=0.10.0

# Audio processing
pyaudio>=0.2.13
pygame>=2.5.0
sounddevice>=0.4.6
scipy>=1.11.0

# 3D visualization
pygame>=2.5.0
pyopengl>=3.1.6
pyopengl-accelerate>=3.1.6

# Utilities
python-dotenv>=1.0.0

# Reaper Integration (Optional)
mido>=1.2.10
python-rtmidi>=1.5.0
python-osc>=1.8.0

# Sound Library (Optional)
soundfile>=0.12.0

